refactor: 쿼리 구조 변경 및 뮤테이션 파일 추가
