+++ apps/client/src/pages/my-history/hooks/use-add-performance-mutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { postAddPerformanceToSetList } from '@shared/apis/my-history/setlist-queries';
+import { postAddPerformanceToSetList } from '@shared/apis/my-history/setlist-mutation';
 import { SETLIST_QUERY_KEY } from '@shared/constants/query-key';
 import { SetListPerformance } from '@shared/types/my-history-response';
