+++ apps/client/src/shared/apis/onboard/onboard-mutation.ts
@@ -1,0 +1,12 @@
+import { post } from '@shared/apis/config/instance';
+import { END_POINT } from '@shared/constants/api';
+import { BaseResponse } from '@shared/types/api';
+
+export const postAuthOnboarding = async (
+  favoriteArtists: string[],
+): Promise<void> => {
+  const requestBody = {
+    favoriteArtists: favoriteArtists.map((id) => ({ artistId: id })),
+  };
+  await post<BaseResponse<null>>(END_POINT.POST_AUTH_ONBOARD, requestBody);
+};
